# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Room {
  /** @pattern ^[A-Z0-9]{5}$ */
  code: string;
  createdAt: number;
  /** @minimum 0 */
  participantCount: number;
}

export interface Error {
  error: string;
}