# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Room {
  /** @pattern ^[A-Z0-9]{5}$ */
  code: string;
  createdAt: bigint;
  /** @minimum 0 */
  participantCount: number;
}